backend/pyproject.toml
[project]
name = "aeye-backend"
version = "0.1.0"
description = "Real-time assistive vision backend for Aeye"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.9.0",
    "ultralytics>=8.1.0",
    "easyocr>=1.7.1",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]
